AVRCP: Fix for BMW carkit metadata issue
Fix for remote device not displaying avrcp metadata when user
selectes random song in playlist. If the requested track id is
same as the one that was sent from TG when track was changed and
if the track id is not present in now playing playlist then
send current playing track's metadata.

Change-Id: I3898a5bb7179af58e5a9524a320ff95c5fdd773a